% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_active_compounds.R
\name{get_active_compounds}
\alias{get_active_compounds}
\title{get_active_compounds}
\usage{
get_active_compounds(
  pool,
  herb_ingredient_table,
  ingredient_table,
  herbs,
  ob = NULL,
  dl = NULL,
  ro5 = NULL
)
}
\arguments{
\item{pool}{dbPool}

\item{herbs}{herbs name chars}

\item{ob}{OB thr}

\item{dl}{DL thr}

\item{ro5}{RO5 thr}

\item{table}{table in database}
}
\value{
data
}
\description{
get_active_compounds
}
