% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_disease_targets.R
\name{get_disease_targets}
\alias{get_disease_targets}
\title{Retrieve the targets of disease from the data table}
\usage{
get_disease_targets(pool, disease_target_table, disease)
}
\arguments{
\item{pool}{dbPool}

\item{disease_target_table}{disease_target table name in database}

\item{disease}{disease name}
}
\value{
data
}
\description{
Retrieve the targets of disease from the data table
}
\examples{
library(pool)
library(DBI)
library(RSQLite)

disease <- "chronic heart failure"
dbname <- system.file("app/data/tcmie.db", package = "TCMIE")

pool <- pool::dbPool(
  drv = RSQLite::SQLite(),  # SQLite 驱动
  dbname = db
)

targets <- get_disease_targets(pool,"disease_target",disease)
}
