% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app_config.R
\name{init_db_pool}
\alias{init_db_pool}
\title{Initialize SQLite database connection pool}
\usage{
init_db_pool()
}
\value{
A database connection pool object.
}
\description{
Initialize SQLite database connection pool
}
